Commit partial code
