+++ evo_test_manager.py
@@ -1,4 +1,5 @@
 import argparse
+import random
 
 from rocket_controller.cli_helper import process_args, str_to_strategy
 from rocket_controller.packet_server import serve
@@ -7,9 +8,10 @@ from rocket_controller.strategies import Strategy
 
 class EvoTestManager:
 
-    def run(self, encoding):
+    @staticmethod
+    def run(encoding):
         params_dict = process_args(
-            argparse.Namespace(strategy="EvoPriorityStrategy",
+            argparse.Namespace(strategy="EvoDelayStrategy",
                                nodes=3,
                                partition=None,
                                nodes_unl=None,
@@ -17,10 +19,10 @@ class EvoTestManager:
                                config=None,
                                overrides={'encoding': encoding})
         )
-        strategy: Strategy = str_to_strategy('RandomFuzzer')(**params_dict)
+        strategy: Strategy = str_to_strategy('EvoDelayStrategy')(**params_dict)
         server = serve(strategy)
         server.wait_for_termination()
 
 
 if __name__ == "__main__":
-    EvoTestManager.run([])
+    EvoTestManager.run([random.randint(0, 4000) for _ in range(42)])
